requirements.txt
numpy==1.21.5
#PyYAML==5.4.1
PyYAML==6.0.2
torch==2.7.1
torchvision==0.22.1
